Create flake for NixOS configuration
Initialized a Nix flake with `nix flake init` and configured a flake
output that builds a NixOS system from the `configuration.nix` in the
git repository. One of the benefits of using flakes is that dependencies
are pinned, which makes building configurations more hermetic.

Upgrading my system to use the latest Nixpkgs is a matter of running
`nix flake update` to recreate the lock file with updated inputs and
then running `nixos-rebuild switch --flake .#shuriken` to build and
activate the new configuration.