Add NixOS module for secrets management
This module allows for secrets management in NixOS configs in a way that
secrets are not stored in plain text when copied into the Nix store when
the flake is evaluated. Secrets are GPG-encrypted at rest inside the Nix
store, and then decrypted and copied to a specified target path upon
system activation via a systemd service.