Initialize Cabal project + NixOS module for tool
Starting work on a tool to abstract some common workflows for NixOS and
dotfiles management:

* Upgrading Nix flake inputs
* Rebuilding and switching NixOS configurations
* Managing symlinks in ~ to files in the Nix store
* Running the Nix store garbage collector

A NixOS module was created to provide a declarative way to specify which
symbolic links to create in the home directory point to different files
in the Nix store. The idea is that module options will be read during
system activation as input to the dot tool for automatic dotfiles
management.

Eventually, these module options will be exposed as flake outputs so
symlinks can be synced ad hoc using the dot tool. This would be useful
for quickly testing and applying new configuration without having to
fully rebuild the NixOS system configuration.